requirements.txt
flask
g4f
pywebview
